refactor:#172 home의 constant 및 type 이동
