+++ src/app/(home)/components/home-ai-feedback-text.tsx
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import { NUTRITION_FEEDBACKS } from '@/constants/home.constant';
+import { NUTRITION_FEEDBACKS } from '@/app/(home)/constants/ai-feedback.constant';
 import { MealNutrition, NutritionGoal } from '@/types/nutrition.type';
 import { calculateNutritionDifferences, getMostSignificantDiff } from '@/app/(home)/utils/nutrition-diff.util';
 import { cn } from '@/lib/utils';
