+++ src/app/(home)/components/home-calendar-month-item.tsx
@@ -1,4 +1,4 @@
-import { MAX_WEEK } from '@/constants/calendar.constant';
+import { MAX_WEEK } from '@/app/(home)/constants/calendar.constant';
 import { formatDateWithDash } from '@/lib/utils/date.util';
 import HomeCalendarWeekItem from './home-calendar-week-item';
