+++ src/app/(home)/components/home-calories-summary-card.tsx
@@ -1,4 +1,4 @@
-import { CIRCUMFERENCE, RADIUS } from '@/constants/home.constant';
+import { CIRCUMFERENCE, RADIUS } from '@/app/(home)/constants/ai-feedback.constant';
 import { formatNumberWithComma } from '@/lib/utils/format-number-with-comma';
 import { getPercentage } from '@/lib/utils/nutrition-calculator.util';
 
@@ -21,9 +21,9 @@ const HomeCaloriesSummaryCard = ({ total, goal }: HomeCalorieSummaryCardProps) =
   return (
     <div className="flex items-center justify-between rounded-2xl bg-white p-4">
       <div>
-        <h3 className="text-gray-550 mb-1 text-[13px]">총 섭취 칼로리</h3>
+        <h3 className="mb-1 text-[13px] text-gray-550">총 섭취 칼로리</h3>
         <span className="mb-2 block text-4xl font-bold">{formatNumberWithComma(total)}</span>
-        <p className="text-gray-550 mb-1 text-sm">
+        <p className="mb-1 text-sm text-gray-550">
           1일 목표 칼로리<span className="ml-1 font-semibold">{formatNumberWithComma(goal)}</span>
         </p>
         <p className="text-sm font-medium">{handleFeedbackMessage()}</p>
