+++ src/app/(home)/types/diff.type.ts
